+++ AES.UpGram.Model/Entity/BaseEntity.cs
@@ -6,7 +6,7 @@ namespace MeConecta.Gram.Domain.Entity
     public abstract class BaseEntity
     {
         [Key]
-        public virtual int Id { get; set; }
+        public virtual long Id { get; set; }
         public DateTime CreatedOn { get; set; } = DateTime.Now;
         public DateTime? UpdatedOn { get; set; }
     }
